CORE: fixed warning in type conversion

+++ src/modules/core/Var.cpp
@@ -78,7 +78,7 @@ void Var::addValueToHistory(const std::string& value) {
 	Value v;
 	v._value = value;
 	v._intValue = string::toInt(v._value);
-	v._longValue = string::toLong(v._value);
+	v._longValue = (long)string::toLong(v._value);
 	v._floatValue = string::toFloat(v._value);
 	_history.push_back(v);
 	Log::debug("new value for %s is %s", _name.c_str(), value.c_str());
